Relocate model utilities to top-level src
