+++ src/Readme.md
@@ -1,5 +1,5 @@
 ## Instruction for code folder
 
-- All classes and functions need to be in component folder
+- Shared classes and functions should live in the top-level `src/` directory.
 - Main loop code needs to be in the main loop code folder.
 - All codes need to have docstrings and hints.
